Telegram Bot/requirements.txt
python-telegram-bot==21.6
PyYAML==6.0.2
pydantic==2.8.2
aiohttp==3.8.4
python-dotenv==1.0.0
playwright==1.47.0
tenacity==9.0.0
yarl==1.9.4
